Create assh.bat
Where:

YOUR_FRIENDLY_HOST_NAME is the name of your host written in the way you want to show in the list for selection. (e.g. Production Server 3 instead of prod-wstd-srv06 or the IP address of prod-wstd-srv06)

USERNAME@YOUR_ACTUAL_HOST_NAME_OR_IP is the actual Partially Qualified Domain Name (PQDN) or IP address of the host you are connecting to.